+++ CHANGELOG.md
@@ -87,7 +87,7 @@
 
 ## 0.2.2
 
-- 新增 isUrl, isVerifyCode
+- 新增 isUrl, isVerifyCode, isBankCardNumber, isIdentityCardNumber
 - 新增 timeToTimeField, timeFieldToTime
 - 新增 hexToRgbaObject, hexToRgbaString
 - 新增 getStringLength, trimString, sliceString, randomString, renderStringTemplate